+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 -[#408](https://github.com/okta/okta-auth-js/pull/408) Provides a polyfill for IE 11+
 -[#410](https://github.com/okta/okta-auth-js/pull/410) Add `token.isLoginRedirect` function to prevent app from starting new Oauth flow while already in OAuth callback state.
 
+### Bug Fixes
+
+-[#420](https://github.com/okta/okta-auth-js/pull/420) Clear accessToken from storage when revoke happen
+
 ## 3.1.4
 
 ### Bug Fixes